refactor: move "internalpayerbackend" which is not a backend to src/audio
